+++ src/main/java/com/berkayb/soundconnect/modules/instrument/entity/Instrument.java
@@ -1,25 +1,24 @@
 package com.berkayb.soundconnect.modules.instrument.entity;
 
 import com.berkayb.soundconnect.shared.entity.BaseEntity;
-import com.berkayb.soundconnect.modules.user.entity.User;
-import jakarta.persistence.*;
-import lombok.*;
-import lombok.experimental.SuperBuilder;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
-@EqualsAndHashCode(callSuper = true)
 @Data
-@SuperBuilder
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
 @Table(name = "tbl_instrument")
 public class Instrument extends BaseEntity {
 	
-	@Column(unique = true, nullable = false)
+	@Column(nullable = false, unique = true)
 	private String name;
 	
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "user_id", nullable = false)
-	private User user;
 	
 }