+++ src/main/java/com/berkayb/soundconnect/modules/instrument/service/IInstrumentService.java
@@ -1,8 +1,0 @@
-package com.berkayb.soundconnect.modules.instrument.service;
-
-import com.berkayb.soundconnect.modules.instrument.dto.request.InstrumentSaveRequestDto;
-import com.berkayb.soundconnect.modules.instrument.entity.Instrument;
-
-public interface IInstrumentService {
-	Instrument saveInstrument(InstrumentSaveRequestDto instrument);
-}