+++ src/main/java/com/berkayb/soundconnect/modules/profile/dto/request/VenueProfileSaveRequestDto.java
@@ -1,0 +1,13 @@
+package com.berkayb.soundconnect.modules.profile.dto.request;
+
+
+import java.util.UUID;
+
+public record VenueProfileSaveRequestDto(
+		String bio,
+		String profilePicture,
+		String instagramUrl,
+		String youtubeUrl,
+		String websiteUrl
+) {
+}