+++ src/main/java/com/berkayb/soundconnect/modules/role/enums/RoleEnum.java
@@ -4,6 +4,6 @@ public enum RoleEnum {
 	ROLE_OWNER,
 	ROLE_ADMIN,
 	ROLE_USER,
-	ROLE_VENUE
-	
+	ROLE_VENUE,
+	ROLE_MUSICIAN
 }