+++ src/main/java/com/berkayb/soundconnect/modules/user/dto/request/UserSaveRequestDto.java
@@ -15,7 +15,5 @@ public record UserSaveRequestDto(
 		UUID roleId,
 		City city,
 		String password
-		// FIXME: enstruman endpointi hazır olduğunda instrumentIds kısmı tekrar aktif hale getirilecek
-		// List<UUID> instrumentIds
 ) {
 }