+++ src/main/java/com/berkayb/soundconnect/modules/user/dto/response/UserListDto.java
@@ -1,11 +1,9 @@
 package com.berkayb.soundconnect.modules.user.dto.response;
 
-import com.berkayb.soundconnect.modules.instrument.entity.Instrument;
 import com.berkayb.soundconnect.modules.user.enums.City;
 import com.berkayb.soundconnect.modules.user.enums.Gender;
 import lombok.Builder;
 
-import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 
@@ -15,7 +13,6 @@ public record UserListDto(
 		String username,
 		Gender gender,
 		City city,
-		List<Instrument> instruments,
 		Integer followers,
 		Integer following,
 		Boolean emailVerified,