+++ docs/CHANGELOG.md
@@ -1,16 +1,65 @@
 ### updated:
-- yml
-- AuthProvider
-- build.gradle
-- EndPoints
+- AuthService
+- CustomUserDetailsService
+- Endpoints
+- ErrorType
+- JwtAuthenticationFilter
+- JwtTokenProvider
+- ReqisterRequestDto
 - User
-- UserDetailsImpl
-- UserRepository
+- UserServiceImpl
+- UserStatus
+- UserUpdateRequestDto
+- Venue
+- VenueMapper
+- VenueRepository
+- VenueRequestDto
+- VenueResponseDto
+- VenueServiceImpl
+- 
+- 
+- 
+- 
+- 
 
+### deleted:
+Instrument Module
 
 ### new:
-- GoogleAuthController
-- GoogleAuthControllerImpl
-- GoogleAuthRequestDto
-- GoogleAuthService
-- CHANGELOG.md
+- ArtistVenueConnectionRequest
+- ArtistVenueConnectionRequestController
+- ArtistVenueCOnnectionRequestControllerImpl
+- ArtisVenueConnectionRequestCreateDto
+- ArtisVenueConnectionRequestMapper
+- ArtisVenueConnectionRequestRepository
+- ArtisVenueConnectionRequestResponseDto
+- ArtisVenueConnectionRequestService
+- ArtisVenueConnectionRequestServiceImpl
+- JwtUtil
+- MusicianProfile
+- MusicianProfileController
+- MusicianProfileControllerImpl
+- MusicianProfileCreator
+- MusicianProfileMapper
+- MusicianProfileRepository
+- MusicianProfileResponseDto
+- MusicianProfileSaveRequestDto
+- MusicianProfileService
+- MusicianProfileServiceImpl
+- ProfileCreator
+- ProfileFactory
+- RequestByType
+- RequestStatus
+- VenueProfile
+- VenueProfileController
+- VenueProfileControllerImpl
+- VenueProfileMapper
+- VenueProfileRepository
+- VenueProfileResponseDto
+- VenueProfileSaveRequestDto
+- VenueProfileService
+- VenueProfileServiceImpl
+tekrar pushlicam commit guncellemedim :D 
+- 
+- 
+-